moves files back into their appropriate modules and removes junk lines of code
